chore: restructure code - dashboard views - admin
